requirements.txt
streamlit
fastai2
